chatbot/dj/app/requirements.txt
langchain>=0.3.24
langgraph>=0.3.34
langchain-community>=0.3.22
langchain-google-vertexai>=2.0.21
pydantic>=2.7.1
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.108.0
uvicorn>=0.25.0
redis>=5.0.0
pillow>=10.0.0
google-cloud-aiplatform>=1.90.0
langfuse>=2.60.0 
streamlit>=1.44.0
nest-asyncio==1.6.0
sseclient>=0.0.27
langchain-mcp-adapters>=0.0.9